# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PLUG API
  version: 0.2.1
  description: >
    Phase 0 and Phase 1 contract shared by the PLUG iOS app and the modular-monolith backend.
    Source of truth for method, route, auth, request, response, enums and errors
    (manual.docx §18.1). Changes require one pull request in /contracts with BOTH
    approvals, updated fixtures and an entry in CHANGELOG.md.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-staging.plug.app
    description: Staging
paths:
...
  /v1/auth/apple:
    post:
      operationId: signInWithApple
      summary: Exchange an Apple identity token for a PLUG session
      description: >
        The client sends the identity token Sign in with Apple returned on the device.
        The server verifies the signature against Apple's published keys and checks the
        issuer, audience, expiry and nonce before it trusts anything in the token. A token
        that has already been exchanged is refused, so a captured token cannot be replayed.

        Sending a guest access token in the Authorization header upgrades that guest
        account in place: the user_id does not change, so a request the guest started
        stays attached to the same person after signing in.
      security:
        - plugSession: []
        - {}
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppleSignIn' }
      responses:
        '201':
          description: Signed in. A new session was issued.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/google:
    post:
      operationId: signInWithGoogle
      summary: Exchange an Google identity token for a PLUG session
      description: >
        The client sends the identity token Sign in with Google returned on the device.
        The server verifies the signature against Google's published keys and checks the
        issuer, audience, expiry and nonce before it trusts anything in the token. A token
        that has already been exchanged is refused, so a captured token cannot be replayed.

        Sending a guest access token in the Authorization header upgrades that guest
        account in place: the user_id does not change, so a request the guest started
        stays attached to the same person after signing in.
      security:
        - plugSession: []
        - {}
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoogleSignIn' }
      responses:
        '201':
          description: Signed in. A new session was issued.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/phone/start:
    post:
      operationId: startPhoneVerification
      summary: Begin phone verification and send a one-time code
      description: >
        Returns a challenge the caller verifies with POST /v1/auth/phone/verify. The
        response is identical whether or not the number is already known to PLUG, so
        this endpoint cannot be used to discover who has an account.

        Three independent limits apply: per phone number, per caller address and, on
        verification, per challenge. The response never says which limit was reached.

        Twilio SMS delivery is enabled only when explicitly configured;
        environments without a delivery channel return 503 dependency_unavailable
        rather than pretending a message was sent.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhoneStart' }
      responses:
        '202':
          description: A code was sent if the number can receive one.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhoneChallenge' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/phone/verify:
    post:
      operationId: verifyPhoneCode
      summary: Exchange a verified phone code for a PLUG session
      description: >
        A wrong or expired code is a validation_failed carrying details[0].field = "code"
        and details[0].code = "invalid" or "expired". The client branches on that token
        and never on the message. Running out of attempts is a rate_limited response with
        retry_after_seconds, because the control that stopped the caller is a limit.

        Sending a guest access token in the Authorization header upgrades that guest
        account in place, exactly as the Apple route does.
      security:
        - plugSession: []
        - {}
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhoneVerify' }
      responses:
        '201':
          description: Verified. A new session was issued.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/guest:
    post:
      operationId: signInAsGuest
      summary: Create a guest account and session
      description: >
        A guest is a real account with a real user_id, so the work a guest starts survives
        an upgrade to Apple or phone. It carries the guest scope only, and its refresh
        window is shorter than a verified account's.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GuestSignIn' }
      responses:
        '201':
          description: A guest account and session were created.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Rotate a refresh token and issue a new access token
      description: >
        Every refresh rotates: the token sent here is spent and a new one comes back.
        Presenting a token that was already rotated means the material leaked, so the
        whole chain is revoked and the caller must sign in again.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: A rotated session.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the calling session server-side
      description: >
        Revocation happens on the server, which is what makes "I logged out" true. The
        access token and the whole refresh chain stop working immediately. Calling this
        with an already revoked token still returns 204, so a client that retries after a
        dropped connection does not see an error for work that is already done.
      security:
        - plugSession: []
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '204':
          description: The session is revoked.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestIdResponseHeader' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    stagingJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staging issuer, audience, expiry and plug.requests.write scope are validated. Local loopback development is exempt.
    plugSession:
      type: http
      scheme: bearer
      description: >
        An opaque PLUG access token from POST /v1/auth/*. It is not a JWT and carries no
        claims: the server looks the session up on every request, which is what lets logout
        and a security-sensitive change revoke it the instant they happen rather than when
        an expiry eventually catches up (manual.docx 25.2).
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      schema: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
      description: >
        Client-generated correlation ID. Echoed back on the response and carried into
        both lanes' structured logs so the connected checkpoint can find one request
        in both places. The server generates one if absent or invalid (manual.docx §19.2).
  headers:
    RequestIdResponseHeader:
      description: The correlation ID for this request — either the caller's or a server-generated one.
      schema: { type: string }
  responses:
    ValidationFailed:
      description: The request was malformed or out of range.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthenticated:
      description: Missing or invalid staging access token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Conflict:
      description: The identity is already attached to a different account.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    DependencyUnavailable:
      description: A dependency this route needs is not available.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Too many requests; retry_after_seconds names the wait
      headers:
        Retry-After:
          schema: { type: string }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    InternalError:
      description: Unexpected internal failure; no internal details are exposed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    AppleSignIn:
      type: object
      additionalProperties: false
      required: [identity_token, nonce, consent_version]
      properties:
        identity_token:
          type: string
          minLength: 1
          maxLength: 4096
          description: The identity token from ASAuthorizationAppleIDCredential. Never logged.
        nonce:
          type: string
          minLength: 16
          maxLength: 128
          pattern: '^[A-Za-z0-9_-]+$'
          description: >
            The raw nonce whose SHA-256 the client sent to Apple. The server hashes it and
            compares it to the token's nonce claim, which is what ties the token to this
            device's sign-in attempt rather than to any token Apple ever issued.
        consent_version: { $ref: '#/components/schemas/ConsentVersion' }
        intent: { $ref: '#/components/schemas/AuthIntent' }
    GoogleSignIn:
      type: object
      additionalProperties: false
      required: [identity_token, nonce, consent_version]
      properties:
        identity_token:
          type: string
          minLength: 1
          maxLength: 4096
          description: The identity token from GIDSignInResult.user.idToken. Never logged.
        nonce:
          type: string
          minLength: 16
          maxLength: 128
          pattern: '^[A-Za-z0-9_-]+$'
          description: >
            The raw nonce sent to the Google SDK. It must match the signed token nonce exactly.
        consent_version: { $ref: '#/components/schemas/ConsentVersion' }
        intent: { $ref: '#/components/schemas/AuthIntent' }
    PhoneStart:
      type: object
      additionalProperties: false
      required: [phone_number]
      properties:
        phone_number:
          type: string
          pattern: '^\+[1-9][0-9]{7,14}$'
          description: E.164, including the country code. Stored as a hash, never logged.
    PhoneChallenge:
      type: object
      additionalProperties: false
      required: [challenge_id, expires_at, attempts_remaining]
      properties:
        challenge_id: { type: string, pattern: '^cha_[A-Za-z0-9-]+$' }
        expires_at: { type: string, format: date-time }
        attempts_remaining: { type: integer, minimum: 0, maximum: 5 }
    PhoneVerify:
      type: object
      additionalProperties: false
      required: [challenge_id, code, consent_version]
      properties:
        challenge_id: { type: string, pattern: '^cha_[A-Za-z0-9-]+$' }
        code:
          type: string
          pattern: '^[0-9]{6}$'
          description: The six digits from the message. Never logged, in any form.
        consent_version: { $ref: '#/components/schemas/ConsentVersion' }
        intent: { $ref: '#/components/schemas/AuthIntent' }
    AuthIntent:
      type: string
      enum: [sign_up, sign_in]
      description: >
        sign_up refuses an existing verified provider identity (409 conflict, detail code
        account_exists); sign_in refuses an unknown verified identity (account_not_found).
        These checks run only after provider token or OTP verification. Omission preserves
        the combined flow for older clients. Email strings never link different providers.
    GuestSignIn:
      type: object
      additionalProperties: false
      required: [consent_version]
      properties:
        consent_version: { $ref: '#/components/schemas/ConsentVersion' }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 1, maxLength: 512 }
...
    ConsentVersion:
      type: string
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
      example: "2026-09-01"
      description: >
        The publication date of the Terms and Privacy Policy the person accepted. The
        client sends the version it actually displayed, and the server rejects a version
        it has never published rather than recording an agreement to unknown text.
    Session:
      description: >
        Issued by every sign-in route and by refresh. Both tokens are opaque and both are
        stored only in the iOS Keychain (manual.docx 20.2). Access tokens live 15 minutes;
        a refresh token lives 30 days for a verified account and 7 days for a guest.
      type: object
      additionalProperties: false
      required: [access_token, access_token_expires_at, refresh_token, refresh_token_expires_at, account, consent]
      properties:
        access_token: { type: string, minLength: 1, maxLength: 512 }
        access_token_expires_at: { type: string, format: date-time }
        refresh_token: { type: string, minLength: 1, maxLength: 512 }
        refresh_token_expires_at: { type: string, format: date-time }
        account: { $ref: '#/components/schemas/Account' }
        consent: { $ref: '#/components/schemas/Consent' }
    Account:
      type: object
      additionalProperties: false
      required: [user_id, type, scopes]
      properties:
        user_id:
          type: string
          pattern: '^usr_[A-Za-z0-9-]+$'
          description: >
            Stable for the life of the account, including across a guest upgrade. Phase 2
            and everything after it use this identifier; there is no second client identity.
        type: { type: string, enum: [guest, phone, apple, google] }
        scopes:
          type: array
          description: What this session may do. A guest holds guest only.
          items: { type: string, enum: [guest, member, admin] }
    Consent:
      type: object
      additionalProperties: false
      required: [current_version, accepted_version]
      properties:
        current_version: { $ref: '#/components/schemas/ConsentVersion' }
        accepted_version:
          oneOf:
            - { $ref: '#/components/schemas/ConsentVersion' }
            - { type: 'null' }
          description: >
            Null only in the moment before a first acceptance. When it trails
            current_version the client must show the consent screen again.
        accepted_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
...
    Error:
      description: The error envelope. Every non-2xx response has exactly this shape (manual.docx §18.2).
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Stable and machine-readable. Clients branch on this, never on the message.
              enum: [validation_failed, unauthenticated, forbidden, not_found, conflict,
                     rate_limited, restricted_intent, dependency_unavailable, internal_error]
            message:
              type: string
              description: Human-readable and safe to display. Never contains internal identifiers, stack traces, SQL or provider error text.
            request_id: { type: string, example: req_01H8XK4Q7N }
            details:
              type: array
              description: Field-level problems. validation_failed only.
              items:
                type: object
                additionalProperties: false
                required: [field, code]
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
            retry_after_seconds:
              type: integer
              description: Present on rate_limited and dependency_unavailable.